Update to Flux v3
Use a double for the animation timestamp.

+++ flux-ffi/src/flux.h
@@ -7,7 +7,7 @@ typedef struct Flux Flux;
 
 struct Flux *flux_new(float width, float height, float physical_width, float physical_height, const char *settings_json_ptr);
 
-void flux_animate(struct Flux *ptr, float timestamp);
+void flux_animate(struct Flux *ptr, double timestamp);
 
 void flux_resize(struct Flux *ptr, float logical_width, float logical_height, float physical_width, float physical_height);
